simple powershell module to get single or all user/contact certificates from an AD with all related information including metadata
.SYNOPSIS
simple module to get single or all user/contact certificate from an AD
look for a certificate in usercert, usercertificate,
usersmimecertificate attributes for contact and user object
2 functions available : Get-ADUserCertificate and
Get-AllADUserCertificates
the function are standalone and the code could be used outside the
module easily. the only prerequisite is RSAT with AD cmdlets.

.DESCRIPTION
Require RSAT if used on non Domain Controller environment.
Do not manage manual authentication to directory (to be managed in a
future version)

.EXAMPLE
C:\PS> import-module Get-ADUserCertificate.psm1
C:\PS> Remove-Module Get-ADUserCertificate